+++ apps/api-journeys-modern/src/schema/blocks/blocks.zod.ts
@@ -5,6 +5,9 @@ import { ButtonBlockSchema } from './button/button.zod'
 import { CardBlockSchema } from './card/card.zod'
 import { IconBlockSchema } from './icon/icon.zod'
 import { ImageBlockSchema } from './image/image.zod'
+import { RadioOptionBlockSchema } from './radioOption/radioOption.zod'
+import { RadioQuestionBlockSchema } from './radioQuestion/radioQuestion.zod'
+import { SignUpBlockSchema } from './signUpBlock/signUpBlock.zod'
 import { TextResponseBlockSchema } from './textResponse/textResponse.zod'
 import { TypographyBlockSchema } from './typography/typography.zod'
 import { VideoBlockSchema } from './video/video.zod'
@@ -14,22 +17,6 @@ const BlockSchema = z.object({
   journeyId: z.string(),
   parentBlockId: z.string().nullable(),
   parentOrder: z.number().nullable()
-})
-
-// RadioOptionBlock schema
-const RadioOptionBlockSchema = BlockSchema.extend({
-  label: z.string(),
-  action: ActionSchema.nullable()
-})
-
-// RadioQuestionBlock schema
-const RadioQuestionBlockSchema = BlockSchema.extend({})
-
-// SignUpBlock schema
-const SignUpBlockSchema = BlockSchema.extend({
-  action: ActionSchema,
-  submitIconId: z.string().nullable(),
-  submitLabel: z.string().nullable()
 })
 
 // StepBlock schema
