+++ apps/api-journeys-modern/src/schema/blocks/radioQuestion/radioQuestion.zod.ts
@@ -1,0 +1,5 @@
+import { BlockSchema } from '../blocks.zod'
+
+const RadioQuestionBlockSchema = BlockSchema.extend({})
+
+export { RadioQuestionBlockSchema }
