+++ apps/api-journeys-modern/src/schema/blocks/signUpBlock/signUpBlock.zod.ts
@@ -1,0 +1,12 @@
+import { z } from 'zod'
+
+import { ActionSchema } from '../action/action.zod'
+import { BlockSchema } from '../blocks.zod'
+
+const SignUpBlockSchema = BlockSchema.extend({
+  action: ActionSchema,
+  submitIconId: z.string().nullable(),
+  submitLabel: z.string().nullable()
+})
+
+export { SignUpBlockSchema }
